==> scripts/04b-analyzeTopicModels.R <==
##################################################################################################
#
# Analyze Topic Models of EOs
#
##################################################################################################
rm(list = ls())
source("scripts/util/__Util_MASTER.R")

####################
# Analyze coherence by topic number
####################
# Get files
files <- list.files("data_derived/lda_models", full.names = T)

# load each and get coherence
coherence <- lapply(files, function(x) {
  # Load
  print(x)
  load(x)
  # Get coherence data
  to_return <- data.frame(topic_num = length(eo_lda$coherence),
                          coherence = eo_lda$coherence,
                          r2 = eo_lda$r2,
                          logLikelihood = eo_lda$log_likelihood[length(eo_lda$log_likelihood), 2])
  # Return
  return(to_return)
})
model_data <- do.call('rbind', coherence)

# Prep and plot
model_sum <- model_data %>% 
  group_by(topic_num) %>% 
  summarise(mean_cohere = mean(coherence),
            med_cohere = median(coherence),
            r2 = mean(r2),
            logLikelihood = mean(logLikelihood))

gg_cohere <- ggplot() +
  # geom_point(data = model_data, aes(x = topic_num, y = coherence), size = 0.1, color = "grey80") +
  geom_line(data = model_sum, aes(x = topic_num, y = mean_cohere)) +
  geom_point(data = model_sum, aes(x = topic_num, y = r2)) +
  theme_ctokita()
gg_cohere

####################
# Interpret model
####################
rm(list = ls())

# Load bet fit model
load("data_derived/dtms/eo_dtm.Rdata")
load("data_derived/lda_models/eo_lda_k40.Rdata")

plot(eo_lda$log_likelihood, type = "l")

# Get top terms and label topics
eo_lda$top_terms <- GetTopTerms(phi = eo_lda$phi, M = 10)
eo_lda$labels <- LabelTopics(assignments = eo_lda$theta > 0.1, 
                             dtm = eo_dtm,
                             M = 3)

# Summarise 
eo_lda$prevalence <- colSums(eo_lda$theta) / sum(eo_lda$theta) * 100
eo_lda$summary <- data.frame(topic = rownames(eo_lda$phi),
                             label = eo_lda$labels,
                             coherence = round(eo_lda$coherence, 3),
                             prevalence = round(eo_lda$prevalence,3),
                             top_terms = apply(eo_lda$top_terms, 2, function(x){
                               paste(x, collapse = ", ")
                             }),
                             stringsAsFactors = FALSE)
